src/django_project_templates/templates/newsapps_project/requirements.txt
fabric==0.9.0
-e svn+http://code.djangoproject.com/svn/django/trunk#egg=django
-e svn+http://geopy.googlecode.com/svn/tags/release-0.93#egg=geopy
django-debug-toolbar
psycopg2
python-memcached